Modularize tests
Add BrowserFactory to create browser instances
Move method to generate a dimension map to TestMatrix.
Move tests steps (interaction with the website) to functions
Add browserSpy function with screenshot to check out if we're on the
right platform.
Use TestMatrix to iterate over test cases (applying a concurrency limit)
Adapt CapabilityFactory to handle special cases of `platform` vs
`platformName` for the different browsers.