Add type guard for screenshot message
Convert serializer checks to type guards. This makes the guards
reusable for messages.

Change tests command in package.json. Currently, the legacy javaScript
tests fail and I did not manage to exclude them with mocha params. The
glob also seems to be broken, so I have to run two commands, one for the
top-level tests and one for the test in subdirectories.