Add docker-compose environment
Add Dockerfile and docker-compose.yml

Change EnvironmentConfig to allow for partical configuration
(metadata_worker does not need Testingbot credentials).

Add ampq-connection-manager library to wrap connection. The library
automatically reconnects and retries when the connection is lost. This
helps with the start of the workers in docker-compose, where RabbitMQ is
not ready for connections when the workers have started.