Adapt README for Docker-first usage
In production, it's easier to run all commands in the Docker
environment. This change shows how to queue screenshots in the Docker
environment, how to mount the screenshot and config directory with
override files and how to do local development without Docker.

This is for https://phabricator.wikimedia.org/T295061