Use overwrite docker-compose files
Splitting into production and dev parts made it necessary to move the
volume mount definitions out of the docker-compose file and into an
override file. It also allows to keep the RabbitMQ port internal by
default and only expose it in the dev environment.

Document new way of running docker-compose commands