Adapt code for Docker-first usage (#266)
## Adapt code for Docker-first usage

In production, it's easier to run all commands in the Docker
environment. This change updates the README to explain how to queue screenshots in the Docker
environment, how to mount the screenshot and config directory with
override files and how to do local development without Docker.

This is for https://phabricator.wikimedia.org/T295061

## Use overwrite docker-compose files

Splitting into production and dev parts made it necessary to move the
volume mount definitions out of the docker-compose file and into an
override file. It also allows to keep the RabbitMQ port internal by
default and only expose it in the dev environment.

Document new way of running docker-compose commands

## Add Makefile

Add Makefile to make the README more readable and start commands
faster.


Co-authored-by: Abban Dunne <Abban@users.noreply.github.com>