Make metadata summary more robust
Don't throw an exception when the worker can't parse the metadata file
for a campaign. Instead, only process valid files and log the file names
of the invalid files.

This will avoid running the metadata worker in an endless loop where the
error quits the worker, docker-compose restarts the worker and RabbitMQ
re-sends the summary command to the newly started worker.