Fix date parsing of old metadata files
Previous iterations of the metadata generator serialized a `Date` object
the integer value (calling `getTime()`) instead of an ISO-8601 formatted
string. This commit allows for the old numeric values in the metadata.

The `Date` constructor allows both for ISO-formatted strings *and*
numbers, so there is no need to change the unserialization code.