Improve test case update
The test function does a lot of test state updates that are not captured
by the metadata worker and not logged. This commit extends the screenshot
worker to log all the state changes and push a metadata update to the
queue whenever there is a state transition.

This is for https://phabricator.wikimedia.org/T315643