Catch failing browser instantiation
Catch errors when the webdriver API fails (e.g. timeout for mobile
platforms). Set test case state as failed and log.

This also helps triggering the metadata summary in case of errors, since
that is only triggered when *all* tests are no longer pending.

This is for https://phabricator.wikimedia.org/T315643