Give development queue a more meaningful name
This will help identifying where the container belongs to, when listing
containers with `docker ps` or `docker ps -a` (in case of port conflicts
or name conflicts). I decided to keep the name as a meaningful
designator instead of using a Docker-generated one, to help with
identifying the image when listing containers.